Working build-from-scratch
With some remaining problems of course, but I'm happy to declare that
the prototype works!

Replaced unpack.py with build.py

Several things are now created as they should be, and thus don't need
their config checked.  Expanding this set as far as possible is
basically the remaining goal:  get "base" as small as possible, and then
under version control.