Install other files
PyLNP, TwbT, and the various others.  This is so tightly coupled to the
exact config I'm using it would be insane for production code... but
here the whole point is to go fast.

This will break sometime fairly soon, but I can fix it then. Other work,
such as reproducible creation of settings/keybinds/etc, is higher
priority than making this robust.