Refactor build.py; extract.py is separate file
Despite increasing the total amount of code, I think this is easier to
deal with.  Better separation of powers, a clear purpose for each
module, easy to see where further work is needed, etc.

Most importantly, this substantially simplifies the call graph and
reduces the odds of calling functions in the wrong order.